+++ src/validate/configure_main_branch.go
@@ -1,0 +1,35 @@
+package validate
+
+import (
+	"fmt"
+
+	"github.com/fatih/color"
+	"github.com/git-town/git-town/v7/src/dialog"
+	"github.com/git-town/git-town/v7/src/git"
+)
+
+func ConfigureMainBranch(repo *git.ProdRepo) error {
+	localBranches, err := repo.Silent.LocalBranches()
+	if err != nil {
+		return err
+	}
+	oldMainBranch := repo.Config.MainBranch()
+	newMainBranch, err := dialog.Select(dialog.SelectArgs{
+		Options: localBranches,
+		Message: mainBranchPrompt(oldMainBranch),
+		Default: oldMainBranch,
+	})
+	if err != nil {
+		return err
+	}
+	return repo.Config.SetMainBranch(newMainBranch)
+}
+
+func mainBranchPrompt(mainBranch string) string {
+	result := "Please specify the main development branch:"
+	if mainBranch != "" {
+		coloredBranch := color.New(color.Bold).Add(color.FgCyan).Sprintf(mainBranch)
+		result += fmt.Sprintf(" (current value: %s)", coloredBranch)
+	}
+	return result
+}
